move errors to /errors dir
