+++ packages/algob/src/internal/cli/cli.ts
@@ -7,13 +7,13 @@ import debug from "debug";
 import semver from "semver";
 
 import { TASK_HELP, TASK_INIT } from "../../builtin-tasks/task-names";
+import { BuilderError, BuilderPluginError } from "../../errors/errors";
+import { ERRORS } from "../../errors/errors-list";
 import { checkAlgorandUnauthorized } from "../../lib/exceptions";
 import { RuntimeArgs, RuntimeEnv, TaskArguments } from "../../types";
 import { ALGOB_NAME } from "../constants";
 import { BuilderContext } from "../context";
 import { loadConfigAndTasks } from "../core/config/config-loading";
-import { BuilderError, BuilderPluginError } from "../core/errors";
-import { ERRORS } from "../core/errors-list";
 import {
   ALGOB_PARAM_DEFINITIONS,
   ALGOB_SHORT_PARAM_SUBSTITUTIONS
