+++ packages/algob/src/internal/core/tasks/task-definitions.ts
@@ -1,3 +1,5 @@
+import { BuilderError } from "../../../errors/errors";
+import { ErrorDescriptor, ERRORS } from "../../../errors/errors-list";
 import {
   ActionType,
   ParamDefinitionAny,
@@ -5,8 +7,6 @@ import {
   TaskArguments,
   TaskDefinition
 } from "../../../types";
-import { BuilderError } from "../errors";
-import { ErrorDescriptor, ERRORS } from "../errors-list";
 import * as types from "../params/argument-types";
 import { ALGOB_PARAM_DEFINITIONS } from "../params/builder-params";
