+++ packages/algob/test/helpers/environment.ts
@@ -1,9 +1,9 @@
 import debug from "debug";
 
+import { BuilderError } from "../../src/errors/errors";
+import { ERRORS } from "../../src/errors/errors-list";
 import { BuilderContext } from "../../src/internal/context";
 import { loadConfigAndTasks } from "../../src/internal/core/config/config-loading";
-import { BuilderError } from "../../src/internal/core/errors";
-import { ERRORS } from "../../src/internal/core/errors-list";
 import { ALGOB_PARAM_DEFINITIONS } from "../../src/internal/core/params/builder-params";
 import { getEnvRuntimeArgs } from "../../src/internal/core/params/env-variables";
 import { Environment } from "../../src/internal/core/runtime-environment";
