+++ packages/algob/test/internal/core/params/argument-types.ts
@@ -4,7 +4,7 @@ import * as fsExtra from "fs-extra";
 import * as os from "os";
 import * as path from "path";
 
-import { ERRORS } from "../../../../src/internal/core/errors-list";
+import { ERRORS } from "../../../../src/errors/errors-list";
 import * as types from "../../../../src/internal/core/params/argument-types";
 import { expectBuilderError } from "../../../helpers/errors";
